Renaming entrypoints to index
