+++ languages/javascript/stopwatchy-browser/src/index.ts
