+++ languages/javascript/stopwatchy-node/test/getStopWatchFactory.test.ts
@@ -1,4 +1,4 @@
-import { getStopWatchFactory } from "../src/node_entry_point";
+import { getStopWatchFactory } from "../src/index";
 import { NodeStopWatchFactoryImpl } from "../src/implementations/NodeStopWatchFactoryImpl";
 
 test("should produce an instance of StopWatchFactoryImpl", () => {
